feat(winners): implement winners logic
