--- src/main/kotlin/racingcar/domain/Cars.kt
@@ -1,10 +1,0 @@
-package racingcar.domain
-
-class Cars (private val cars:List<Car>){
-
-    companion object{
-        fun generate(names:List<String>):Cars{
-            return Cars(names.map { Car(it.trim()) })
-        }
-    }
-}
