+++ src/main/kotlin/racingcar/model/Cars.kt
@@ -1,0 +1,27 @@
+package racingcar.model
+
+class Cars (private val cars:List<Car>){
+
+    companion object{
+        fun generate(names:List<String>):Cars{
+            return Cars(names.map { Car(it.trim()) })
+        }
+    }
+
+    fun moveAll(){
+        cars.forEach{it.move()}
+    }
+
+    fun findWinners(): List<Car>{
+        val winners = mutableListOf<Car>()
+        val maxPosition = cars.maxOf { it.getPosition() }
+        for(car in cars){
+            if(car.getPosition()==maxPosition){
+                winners.add(car)
+            }
+        }
+        return winners
+    }
+
+
+}
