+++ src/main/kotlin/racingcar/service/CarRace.kt
@@ -1,0 +1,9 @@
+package racingcar.service
+
+import racingcar.model.Cars
+
+class CarRace(private val cars : Cars){
+    fun play(){
+        cars.moveAll()
+    }
+}
